A demo to test the deleting conversation function
Hi All,

I think I found a issue about deleting the conversation local, could you run the demo to check it? Thanks.

You can do this as the following step after you install the app on your device and simulator:
1, Shake your device so that the conversation has a colour meta data.
2, Send a message from Device to Simulator, You will find the meta data is existing on the Simulator side.
3, Click the clear button on the Simulator side.
4, Send a message from Device to Simulator again, you will see the meta data is gone on the Simulator side, but it is existing on the device side.

This a same conversation between the Simulator and Device. I think it is a bug.

Could you check it? Thanks.